+++ scripts/metadata_comparison/test/test_lib_argument_regex.py
@@ -1,9 +1,10 @@
 #!/usr/bin/env python3
 
 import unittest
-from metadata_comparison.lib.argument_regex import *
-from metadata_comparison.extractor import *
 import json
+from metadata_comparison.lib.argument_regex import *
+from metadata_comparison.lib.operation_ids import *
+from metadata_comparison.extractor import find_operation_ids_in_metadata
 
 class ExtractorTestMethods(unittest.TestCase):
